functions/requirements.txt
cloudevents
functions-framework
google-cloud-firestore
